+++ server/src/main/java/com/inventage/portal/gateway/proxy/middleware/languageCookie/LanguageCookieMiddlewareFactory.java
@@ -24,6 +24,7 @@ public class LanguageCookieMiddlewareFactory implements MiddlewareFactory {
     public Future<Middleware> create(Vertx vertx, String name, Router router, JsonObject middlewareConfig) {
         LOGGER.debug("Created '{}' middleware successfully",
             DynamicConfiguration.MIDDLEWARE_LANGUAGE_COOKIE);
-        return Future.succeededFuture(new LanguageCookieMiddleware(name));
+        final String languageCookieName = middlewareConfig.getString(DynamicConfiguration.MIDDLEWARE_LANGUAGE_COOKIE, DEFAULT_LANGUAGE_COOKIE_NAME);
+        return Future.succeededFuture(new LanguageCookieMiddleware(name, languageCookieName));
     }
 }